Minor adjustments

+++ wire/core/WireLog.php
@@ -267,6 +267,19 @@ class WireLog extends Wire {
 	}
 
 	/**
+	 * Does given log name exist?
+	 * 
+	 * @param string $name
+	 * @return bool
+	 * @since 3.0.176
+	 * 
+	 */
+	public function exists($name) {
+		$filename = $this->getFilename($name);
+		return is_file($filename);
+	}
+
+	/**
 	 * Return the given number of entries from the end of log file
 	 * 
 	 * This method is pagination aware.
@@ -447,8 +460,10 @@ class WireLog extends Wire {
 	 *
 	 */
 	public function delete($name) {
+		if(!$this->exists($name)) return false;
 		$log = $this->getFileLog($name);
-		return $log->delete();
+		if($log) return $log->delete();
+		return false;
 	}
 
 	/**
@@ -463,6 +478,7 @@ class WireLog extends Wire {
 	 * 
 	 */
 	public function prune($name, $days) {
+		if(!$this->exists($name)) return false;
 		$log = $this->getFileLog($name);
 		if($days < 1) throw new WireException("Prune days must be 1 or more"); 
 		$oldestDate = strtotime("-$days DAYS");
